requirements.txt
flask
sqlalchemy
psycopg2-binary
python-dotenv